Data.PostgreSQL.NetworkAddress

Description

This module defines network-address types of PostgreSQL. http://www.postgresql.org/docs/current/static/datatype-net-types.html

Documentation

data NetAddress

Constructors

NetAddress4 !V4HostAddress !Word8

NetAddress6 !V6HostAddress !Word8

data V4HostAddress

Host address type along with IPv4 address string.

Constructors

V4HostAddress !Word8 !Word8 !Word8 !Word8

v4HostAddressOctets :: V4HostAddress -> (Word8, Word8, Word8, Word8)

data V6HostAddress

Host address type along with IPv6 address string. Each Word16 value is host byte order. Host byte order is portable in programs on its own host. Network byte order is only needed, when communicating other hosts.

Constructors

V6HostAddress !Word16 !Word16 !Word16 !Word16 !Word16 !Word16 !Word16 !Word16

v6HostAddressLong :: Word16 -> Word16 -> Word16 -> Word16 -> Word16 -> Word16 -> Word16 -> Word16 -> V6HostAddress

v6HostAddressWords :: V6HostAddress -> (Word16, Word16, Word16, Word16, Word16, Word16, Word16, Word16)

v6HostAddress :: [Word16] -> [Word16] -> Maybe V6HostAddress

v6HostAddressL :: [Word16] -> Maybe V6HostAddress

v6HostAddressR :: [Word16] -> Maybe V6HostAddress

newtype Inet

Constructors

Inet NetAddress

newtype Cidr

Constructors

Cidr NetAddress